pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vendsim-vb2"
version = "0.1.0"
description = "OpenEnv-compatible Vending-Bench 2 simulation environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core==0.2.1",
    "fastmcp",
]

[project.optional-dependencies]
server = ["fastapi>=0.115", "uvicorn>=0.34"]
dev = ["pytest>=8.0", "ruff>=0.11"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["vendsim_vb2*"]
